requirements.txt
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
boto3==1.23.0
botocore==1.26.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.0
fastapi==0.75.2
fastapi-jwt-auth==0.5.0
fastapi-limiter==0.1.5
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
mongoengine==0.24.1
nats-py==2.1.2
orjson==3.6.8
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.4.0rc4
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
starlette-context==0.3.6
typing_extensions==4.2.0
ujson==5.4.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
yapf==0.32.0
